+++ packages/expo-router/src/exports.ts
@@ -71,6 +71,7 @@ export {
 } from './primitives';
 
 export {
+  IsWithinNativeNavigator,
   unstable_createStandardRouterNavigator,
   unstable_integrateWithRouter,
 } from './standard-navigation';
@@ -78,6 +79,7 @@ export type {
   IntegrateWithRouterOptions,
   NavigatorContentProps,
   StandardNavigatorDescriptor,
+  StandardNavigatorEmit,
   StandardNavigatorEventMapBase,
   StandardUseNavigationBuilderOptions,
 } from './standard-navigation';